Standardise naming across the canteen features
